move NewHabit into habits subfolder
